requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
pillow
requests
streamlit
pydantic